requirements.txt
slackclient
requests
defang
validators